+++ django_app/core/templates/core/base.html
@@ -48,7 +48,7 @@
                                 </div>
                             </div>
                             <div class="govuk-button-group">
-                                <form method="post" action="{% url 'report_a_suspected_breach:cookies_consent' %}?banner=True">
+                                <form method="post" action="{% url 'cookies_consent' %}?banner=True">
                                     {% csrf_token %}
                                     <button type="submit" class="govuk-button" data-module="govuk-button"
                                             name="do_you_want_to_accept_analytics_cookies" value="True">
@@ -58,7 +58,7 @@
                                             name="do_you_want_to_accept_analytics_cookies" value="False">
                                         Reject analytics cookies
                                     </button>
-                                    <a class="govuk-link" href="{% url 'report_a_suspected_breach:cookies_consent' %}">View
+                                    <a class="govuk-link" href="{% url 'cookies_consent' %}">View
                                         cookies</a>
                                 </form>
                             </div>
@@ -67,7 +67,7 @@
                 {% endif %}
                 {% if request.session.cookies_set_in_banner %}
                     {% if not request.session.hide_cookies_banner %}
-                        <form method="post" action="{% url 'report_a_suspected_breach:hide_cookies' %}?banner=True">
+                        <form method="post" action="{% url 'hide_cookies' %}?banner=True">
                             {% csrf_token %}
                             {% if request.COOKIES.accepted_ga_cookies == "true" %}
                                 <div class="govuk-cookie-banner" data-nosnippet role="region" aria-label="Cookies on Report a Suspected Breach of Trade Sanctions">
@@ -75,7 +75,7 @@
                                         <div class="govuk-grid-row">
                                             <div class="govuk-grid-column-two-thirds">
                                                 <div class="govuk-cookie-banner__content">
-                                                    <p class="govuk-body">You’ve accepted analytics cookies. You can <a class="govuk-link" href="{% url 'report_a_suspected_breach:cookies_consent' %}?redirect=Yes">change your cookie settings</a> at any time.</p>
+                                                    <p class="govuk-body">You’ve accepted analytics cookies. You can <a class="govuk-link" href="{% url 'cookies_consent' %}?redirect=Yes">change your cookie settings</a> at any time.</p>
                                                 </div>
                                             </div>
                                         </div>
@@ -93,7 +93,7 @@
                                         <div class="govuk-grid-row">
                                             <div class="govuk-grid-column-two-thirds">
                                                 <div class="govuk-cookie-banner__content">
-                                                    <p class="govuk-body">You’ve rejected analytics cookies. You can <a class="govuk-link" href="{% url 'report_a_suspected_breach:cookies_consent' %}?redirect=Yes">change your cookie settings</a> at any time.</p>
+                                                    <p class="govuk-body">You’ve rejected analytics cookies. You can <a class="govuk-link" href="{% url 'cookies_consent' %}?redirect=Yes">change your cookie settings</a> at any time.</p>
                                                 </div>
                                             </div>
                                         </div>
@@ -254,7 +254,7 @@
                     </div>
                     <div class="govuk-footer__meta-item">
                         <a class="govuk-footer__link"
-                           href="{% url 'report_a_suspected_breach:cookies_consent' %}">Cookies</a>
+                           href="{% url 'cookies_consent' %}">Cookies</a>
                     </div>
                     <div class="govuk-footer__meta-item">
                         <a class="govuk-footer__link"
