+++ django_app/core/templates/core/cookies_consent.html
@@ -4,7 +4,7 @@
 {% block form_content %}
   {% block column_content %}
     <h1 class="govuk-heading-xl">Cookies</h1>
-    <p class="govuk-body"><a href="/" class="govuk-link">Report a suspected breach of trade sanctions</a> puts small files (known as ‘cookies’) on your computer.</p>
+    <p class="govuk-body"><a href="/django_app/core/static" class="govuk-link">Report a suspected breach of trade sanctions</a> puts small files (known as ‘cookies’) on your computer.</p>
 
     <p class="govuk-body">These cookies are used across the Report a suspected breach of trade sanctions application.</p>
