+++ django_app/core/urls.py
@@ -1,4 +1,4 @@
-from core.views import RedirectBaseDomainView
+from core.views import CookiesConsentView, HideCookiesView, RedirectBaseDomainView
 from django.conf import settings
 from django.contrib import admin
 from django.urls import include, path
@@ -7,6 +7,8 @@ urlpatterns = [
     path("", RedirectBaseDomainView.as_view(), name="initial_redirect_view"),
     path("report_a_suspected_breach/", include("report_a_suspected_breach.urls")),
     path("view_a_suspected_breach/", include("view_a_suspected_breach.urls")),
+    path("cookies_consent", CookiesConsentView.as_view(), name="cookies_consent"),
+    path("hide_cookies", HideCookiesView.as_view(), name="hide_cookies"),
     path("feedback/", include("feedback.urls")),
     path("pingdom/", include("healthcheck.urls")),
     path("throw_error/", lambda x: 1 / 0),
