+++ django_app/report_a_suspected_breach/views.py
@@ -27,13 +27,7 @@ from utils.s3 import (
 )
 
 from .choices import TypeOfRelationshipChoices
-from .forms import (
-    CookiesConsentForm,
-    EmailVerifyForm,
-    HideCookiesForm,
-    SummaryForm,
-    UploadDocumentsForm,
-)
+from .forms import EmailVerifyForm, SummaryForm, UploadDocumentsForm
 from .models import Breach, PersonOrCompany, ReporterEmailVerification, SanctionsRegime
 from .tasklist import (
     AboutThePersonOrBusinessTask,
@@ -457,76 +451,6 @@ class ReportABreachWizardView(BaseWizardView):
             self.storage.current_step = self.steps.first
 
         return redirect(reverse("report_a_suspected_breach:complete"))
-
-
-class CookiesConsentView(FormView):
-    template_name = "report_a_suspected_breach/cookies_consent.html"
-    form_class = CookiesConsentForm
-
-    def get_form_kwargs(self) -> dict[str, Any]:
-        kwargs = super().get_form_kwargs()
-        kwargs["request"] = self.request
-        initial_dict = {}
-        self.request.session["show_cookies_form"] = True
-
-        # set the referer so the user can click the redirect link to get to their original page
-        # if the referer is the cookies consent page, send the user back to the tasklist
-        if referer := self.request.META.get(
-            "HTTP_REFERER",
-        ):
-            if not reverse("report_a_suspected_breach:cookies_consent") in referer:
-                self.request.session["cookies_referer"] = referer
-                self.request.session.modified = True
-
-        if current_cookies_policy := self.request.COOKIES.get("accepted_ga_cookies"):
-            initial_dict["accept_cookies"] = current_cookies_policy == "true"
-            kwargs["initial"] = initial_dict
-
-        if self.request.GET.get("redirect"):
-            self.request.session["redirected_from_banner"] = True
-            self.request.session.modified = True
-
-        if self.request.session.get("submitted_on_cookies_page"):
-            self.request.session.pop("show_cookies_form")
-
-        return kwargs
-
-    def form_valid(self, form: CookiesConsentForm) -> HttpResponse:
-        # cookie consent lasts for 1 year
-        cookie_max_age = 365 * 24 * 60 * 60
-        response = redirect("/")
-
-        self.request.session.pop("redirected_from_banner", None)
-
-        if self.request.GET.get("banner", ""):
-            self.request.session["cookies_set_in_banner"] = True
-            self.request.session.modified = True
-        else:
-            # the user submitted the form on the cookies_consent page
-            # we want to keep them there after the post
-            self.request.session["submitted_on_cookies_page"] = True
-            response = redirect(reverse("report_a_suspected_breach:cookies_consent"))
-
-        # regardless of their choice, we set a cookie to say they've made a choice
-        response.set_cookie("cookie_preferences_set", "true", max_age=cookie_max_age)
-        response.set_cookie(
-            "accepted_ga_cookies",
-            "true" if form.cleaned_data["do_you_want_to_accept_analytics_cookies"] else "false",
-            max_age=cookie_max_age,
-        )
-
-        return response
-
-
-class HideCookiesView(FormView):
-    template_name = "report_a_suspected_breach/hide_cookies.html"
-    form_class = HideCookiesForm
-
-    def form_valid(self, form: HideCookiesForm) -> HttpResponse:
-        self.request.session["hide_cookies_banner"] = True
-        self.request.session.modified = True
-        referrer_url = self.request.GET.get("referrer_url", "/")
-        return redirect(referrer_url)
 
 
 class CompleteView(TemplateView):
